Move #close? to Numeric
darcs-hash:20071205131854-4fc50-b83e8ad623c548b8ff5c264c8ae915c8e293b83f.gz

+++ lib/bacon.rb
@@ -208,7 +208,7 @@ class Proc
   end
 end
 
-class Float
+class Numeric
   def close?(to, delta)
     (to.to_f - self).abs <= delta.to_f
   rescue
